// src/types/productDetails.types.ts
import { Brand } from "./brand.types"
import { Category, Subcategory } from "./category.types"
import { Review } from "./reveiw.types"

export interface ProductDetailsType {
  sold: number
  images: string[]
  subcategory: Subcategory[]
  ratingsQuantity: number
  _id: string
  title: string
  slug: string
  description: string
  quantity: number
  price: number
  imageCover: string
  category: Category
  brand: Brand
  ratingsAverage: number
  createdAt: string
  updatedAt: string
  __v: number
  reviews: Review[]
  id: string
}